move components to separate directories
