+++ assets/js/main.js
@@ -1,4 +1,4 @@
-/* globals $, moment, Cookies, Draggabilly, Packery, updateGrid */
+/* globals $, moment, Cookies, Draggabilly, Packery, updateGrid, COMPONENTS */
 
 let worldState;
 let updateTime;
@@ -588,198 +588,7 @@ function updateNews() {
   }
 }
 
-function getLabelColor(faction) {
-  switch (faction) {
-  case 'Corpus':
-    return 'label-info';
-  case 'Grineer':
-    return 'label-danger';
-  case 'Infested':
-    return 'label-success';
-  case 'Corrupted':
-    return 'label-warning';
-  default:
-    return 'label-default';
-  }
-}
-
-function getProgressBarColor(faction) {
-  switch (faction) {
-  case 'Corpus':
-    return 'corpus-invasion';
-  case 'Grineer':
-    return 'grineer-invasion';
-  case 'Infested':
-    return 'infested-invasion';
-  case 'Corrupted':
-    return 'corrupted-invasion';
-  default:
-    return 'default-invasion';
-  }
-}
-
-const COMPONENTS = {
-  alerts: {
-    id: 'alerts',
-    worldStateKey: 'alerts',
-    parse(data) {
-      let numAlerts = 0;
-
-      if (platformSwapped) {
-        // this should happen when changing platforms, not here
-        // I just want to change as little code as possible for now
-        this.cleanup();
-      }
-
-      data.forEach(alert => {
-        if (!alert.expired) {
-          numAlerts += 1;
-          if ($(`#${alert.id}`).length === 0) {
-            this.add(alert);
-          }
-        }
-        // if it's expired we don't care, timer will delete this (this.expired)
-      });
-
-      if (numAlerts > 0) {
-        $('#alerttitle').hide();
-      } else {
-        $('#alerttitle').html('No active alerts :(').show();
-      }
-    },
-    add(alert) {
-      const {mission} = alert;
-      const alertElement = cloneTemplate(this.id);
-      alertElement.attr('id', alert.id);
-
-      alertElement.find('.node').html(mission.node);
-      alertElement.find('.mission-type').html(mission.type);
-      alertElement.find('.faction').html(mission.faction);
-      alertElement.find('.levels')
-        .html(`${mission.minEnemyLevel}-${mission.maxEnemyLevel}`);
-      alertElement.find('.credits').html(mission.reward.credits);
-
-      if (mission.archwingRequired) {
-        cloneTemplate('archwing').prependTo(alertElement).after(' ');
-      }
-      if (mission.nightmare) {
-        cloneTemplate('nightmare').prependTo(alertElement).after(' ');
-      }
-
-      const newLine = alertElement.find('br');
-
-      const timer = cloneTimer(alert.activation, alert.expiry, this.id);
-      manageTimer(timer);
-      newLine.before(timer);
-
-      const rewards = parseRewards(mission.reward);
-      for (const item of rewards) {
-        newLine.before(cloneLabel('right', 'label-info', item)
-          .css('margin-right', '5px'));
-      }
-
-      $('#alertbody').before(alertElement);
-    },
-    expired(timer) {
-      timer.closest('li').remove();
-      updateGrid();
-    },
-    cleanup() {
-      $('#alertList').children().not('#alertbody').remove();
-    },
-  },
-  invasions: {
-    id: 'invasions',
-    worldStateKey: 'invasions',
-    parse(data) {
-      let numInvasions = 0;
-
-      if (platformSwapped) {
-        // this should happen when changing platforms, not here
-        // I just want to change as little code as possible for now
-        this.cleanup();
-      }
-
-      data.forEach(invasion => {
-        const invasionElement = $(`#${invasion.id}`);
-        if (invasionElement.length !== 0) {
-          if (invasion.completed) {
-            invasionElement.remove();
-          } else {
-            this.update(invasion, invasionElement);
-            numInvasions += 1;
-          }
-        } else if (!invasion.completed) {
-          this.add(invasion);
-          numInvasions += 1;
-        }
-      });
-
-      if (numInvasions > 0) {
-        $('#invasiontitle').html('*End time is estimated');
-      } else {
-        $('#invasiontitle').html('No active invasions :(');
-      }
-    },
-    add(invasion) {
-      const invasionElement = cloneTemplate(this.id);
-      invasionElement.attr('id', invasion.id);
-
-      invasionElement.find('.node').html(invasion.node);
-      invasionElement.find('.desc').html(invasion.desc);
-
-      const rewardsElement = invasionElement.find('.rewards');
-      ['attack', 'defend'].forEach(side => {
-        const factionName = invasion[`${side}ingFaction`];
-        const labelColor = getLabelColor(factionName);
-        const literalSide = (side === 'attack' ? 'left' : 'right');
-        const rewards = parseRewards(invasion[`${side}erReward`]);
-
-        for (const item of rewards) {
-          cloneLabel(literalSide, labelColor, item)
-            .appendTo(rewardsElement);
-        }
-
-        invasionElement
-          .find(`.${side}`).addClass(getProgressBarColor(factionName))
-          .find('img').attr('src', getFactionPicture(factionName));
-      });
-
-      this.update(invasion, invasionElement);
-
-      $('#invasionbody').before(invasionElement);
-    },
-    update(invasion, invasionElement) {
-      const attackPercent = Math.floor(invasion.vsInfestation
-        ? invasion.completion / 2
-        : invasion.completion);
-      const defendPercent = 100 - attackPercent;
-
-      const attackBar = invasionElement.find('.attack');
-      const defendBar = invasionElement.find('.defend');
-
-      if (invasion.count > 0) {
-        attackBar.addClass('winning-right');
-        defendBar.removeClass('winning-left');
-      } else {
-        attackBar.removeClass('winning-right');
-        defendBar.addClass('winning-left');
-      }
-
-      attackBar.css('width', `${attackPercent}%`)
-        .attr('aria-valuenow', `${attackPercent}%`);
-      defendBar.css('width', `${defendPercent}%`)
-        .attr('aria-valuenow', `${defendPercent}%`);
-
-      invasionElement.find('.eta').html(invasion.eta);
-    },
-    cleanup() {
-      $('#invasionList').children().not('#invasionbody').remove();
-    },
-  },
-};
-
-function parseRewards(rewardsData) {
+function parseRewards(rewardsData) { // eslint-disable-line no-unused-vars
   const rewards = rewardsData.items.slice();
   // const rewardsSet = new Set(rewards);
 
@@ -804,7 +613,7 @@ function cloneLabel(pullDirection, colorClass, text) {
     .addClass(colorClass);
 }
 
-function cloneTimer(start, end, component) {
+function cloneTimer(start, end, component) { // eslint-disable-line no-unused-vars
   return cloneLabel('right').addClass('timer')
     .attr('data-start', (new Date(start)).getTime())
     .attr('data-end', (new Date(end)).getTime())
@@ -866,11 +675,11 @@ function removeTimeBadgeColor(element) {
   element.removeClass('label-info');
 }
 
-// helper function, transforms miliseconds diff into one of:
-// Dd HHh MMm SSs
-// Hh MMm SSs
-// Mm SSs
-// Ss
+/* helper function, transforms miliseconds diff into one of:
+Dd HHh MMm SSs
+Hh MMm SSs
+Mm SSs
+Ss */
 function formatTimer(diff) {
   let timeLeft = diff;
   const stringArray = [];
