+++ components/alerts/template.hbs
@@ -1,0 +1,10 @@
+<li id="alerts-template" class="list-group-item list-group-item-borderless">
+  <b class="node"></b>
+  <br>
+  <div style="margin-top:2px">
+    <b class="mission-type"></b>
+    (<span class="faction"></span>) |
+    <b>Level: </b><span class="levels"></span>
+    | <span class="credits"></span>cr
+  </div>
+</li>
