+++ components/invasions/template.hbs
@@ -1,0 +1,10 @@
+<li id="invasions-template" class="list-group-item list-group-item-borderless" style="padding-bottom:0;">
+  <div class="row text-center"><b class="node"></b><br><span class="desc"></span> (Ends in: <span class="eta"></span>)*</div>
+  <div class="row clearfix rewards" style="margin-bottom: 1px; margin-left:5px; margin-right:5px"></div>
+  <div class="row progress-bar-container" style="margin-bottom: 1px; margin-left:5px; margin-right:5px">
+    <div class="progress">
+      <div class="progress-bar attack" role="progressbar" aria-valuemin="0" aria-valuemax="100"><img class="pull-left faction-invasion-img"></div>
+      <div class="progress-bar defend" role="progressbar" aria-valuemin="0" aria-valuemax="100"><img class="pull-right faction-invasion-img"></div>
+    </div>
+  </div>
+</li>
