+++ package.json
@@ -32,6 +32,7 @@
     "express-handlebars": "^3.0.0",
     "gulp": "^3.9.1",
     "gulp-clean-css": "^3.9.0",
+    "gulp-concat": "^2.6.1",
     "gulp-minify": "^2.1.0",
     "jade": "~1.11.0",
     "morgan": "~1.9.0",
