+++ views/index.hbs
@@ -172,6 +172,7 @@
 <script src="https://unpkg.com/draggabilly@2/dist/draggabilly.pkgd.min.js"></script>
 
 <!-- required js for page -->
+<script src="js/components.js"></script>
 <script src="js/main.js"></script>
