+++ views/partials/templates.hbs
@@ -1,25 +1,6 @@
 <div id="templates" class="hide">
-	<li id="alerts-template" class="list-group-item list-group-item-borderless">
-		<b class="node"></b>
-		<br>
-		<div style="margin-top:2px">
-			<b class="mission-type"></b>
-			(<span class="faction"></span>) |
-			<b>Level: </b><span class="levels"></span>
-			| <span class="credits"></span>cr
-		</div>
-	</li>
-	<li id="invasions-template" class="list-group-item list-group-item-borderless" style="padding-bottom:0;">
-		<div class="row text-center"><b class="node"></b><br><span class="desc"></span> (Ends in: <span class="eta"></span>)*</div>
-		<div class="row clearfix rewards" style="margin-bottom: 1px; margin-left:5px; margin-right:5px"></div>
-		<div class="row progress-bar-container" style="margin-bottom: 1px; margin-left:5px; margin-right:5px">
-			<div class="progress">
-				<div class="progress-bar attack" role="progressbar" aria-valuemin="0" aria-valuemax="100"><img class="pull-left faction-invasion-img"></div>
-				<div class="progress-bar defend" role="progressbar" aria-valuemin="0" aria-valuemax="100"><img class="pull-right faction-invasion-img"></div>
-			</div>
-		</div>
-	</li>
-	<span id="label-template" class="label"></span>
-	<img id="archwing-template" title="Archwing Required for Mission" src="https://i.imgur.com/R1kpRx4.png" class="archwing" height="16px">
-	<img id="nightmare-template" title="Nightmare Mission" src="https://i.imgur.com/x5XoktW.png" class="nightmare" height="16px">
+  {{> concat-templates}}
+  <span id="label-template" class="label"></span>
+  <img id="archwing-template" title="Archwing Required for Mission" src="https://i.imgur.com/R1kpRx4.png" class="archwing" height="16px">
+  <img id="nightmare-template" title="Nightmare Mission" src="https://i.imgur.com/x5XoktW.png" class="nightmare" height="16px">
 </div>
